add seed
