+++ frontend/app/admin/(protected)/accounts/add/page.tsx
@@ -1,27 +1,15 @@
 "use client";
 
-import { useState } from "react";
 import { useRouter } from "next/navigation";
-import Sidebar from "../../../../components/admin-sidebar";
 import Form from "./form";
 import styles from "./page.module.scss";
 
 const AddAdminAccountPage = () => {
-  const [isSidebarOpen, setIsSidebarOpen] = useState(true);
   const router = useRouter();
-
-  const toggleSidebar = () => {
-    setIsSidebarOpen((prev) => !prev);
-  };
 
   return (
     <div className={styles.page_container}>
-      <Sidebar isSideBarOpen={isSidebarOpen} toggleSidebar={toggleSidebar} />
-      <div
-        className={`${styles.main_content} ${
-          isSidebarOpen ? "" : styles.sidebar_collapsed
-        }`}
-      >
+      <div className={`${styles.main_content}`}>
         <div className={styles.page_header}>
           <span className={styles.back_arrow} onClick={() => router.back()}>
             ←
