--- frontend/app/admin/(protected)/accounts/layout.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return (
-    <div className="layout-container">
-
-      <main className="main-content">{children}</main>
-
-    </div>
-  );
-}
